# microservices/contracts/http/jobs.openapi.json
{
  "openapi": "3.1.1",
...
  "info": {
    "title": "JobFind jobs HTTP API",
    "version": "1.0.0",
    "description": "Generated from executable request contracts. JSON only; unknown fields and invalid types are rejected, never silently coerced or removed. Response extensions remain allowed for the legacy migration. Operational endpoints, Socket.IO and legacy /api/* routes are documented separately in docs/http-contracts.md."
  },
  "servers": [
    {
      "url": "/",
      "description": "Direct service origin in the private Compose network; do not publish publicly"
    }
  ],
...
  "paths": {
    "/jobs": {
      "post": {
        "operationId": "jobCreate",
        "tags": [
          "jobs"
        ],
        "summary": "POST /jobs",
        "description": "Modern microservice endpoint; legacy fallback routes are outside this contract. Requires permission job:manage; resource ownership checks still apply. Non-admin callers must belong to an active, approved company. Paid posting requires a current active, approved company even for ADMIN. With an Idempotency-Key, quota, new post, events and original response are committed atomically. Replay returns the initial 201 snapshot, not current moderation state, and rechecks current company membership/approval without spending quota. Keys must not be deleted while retries are possible. A new deadline must be in the future; omitted create deadline defaults to 30 days on the first accepted attempt.",
        "x-jobfind-permission": "job:manage",
        "x-jobfind-internal": false,
        "x-jobfind-company-required": true,
        "x-jobfind-reject-unknown-query": true,
        "security": [
          {
            "internalSecret": []
          }
        ],
        "parameters": [
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            },
            "description": "Optional. Same user + key + canonical input returns the original accepted result; changed input or operation returns 409. Reuse for retries; a new key requests new work."
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]*$"
            },
            "description": "Authenticated identity forwarded by the trusted gateway only."
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ADMIN",
                "COMPANY",
                "EMPLOYER"
              ]
            }
          },
          {
            "name": "x-company-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]*$"
            },
            "description": "Required for non-admin callers; derived from the current account/company store by the gateway."
          },
          {
            "name": "x-company-status",
            "in": "header",
            "required": false,
            "schema": {
              "const": "S1"
            },
            "description": "Required for non-admin callers; derived from the current account/company store by the gateway."
          },
          {
            "name": "x-company-censor",
            "in": "header",
            "required": false,
            "schema": {
              "const": "CS1"
            },
            "description": "Required for non-admin callers; derived from the current account/company store by the gateway."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "headers": {
              "X-Correlation-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errCode": {
                      "const": 0
                    },
                    "data": {
                      "$ref": "#/components/schemas/Job"
                    },
                    "count": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "took": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "emailQueued": {
                      "type": "boolean"
                    },
                    "errMessage": {
                      "type": "string",
                      "maxLength": 1000
                    }
                  },
                  "required": [
                    "errCode",
                    "data"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{id}/repost": {
      "post": {
        "operationId": "jobRepost",
        "tags": [
          "jobs"
        ],
        "summary": "POST /jobs/:id/repost",
        "description": "Modern microservice endpoint; legacy fallback routes are outside this contract. Requires permission job:manage; resource ownership checks still apply. Non-admin callers must belong to an active, approved company. Paid posting requires a current active, approved company even for ADMIN. With an Idempotency-Key, quota, new post, events and original response are committed atomically. Replay returns the initial 201 snapshot, not current moderation state, and rechecks current company membership/approval without spending quota. Keys must not be deleted while retries are possible. A new deadline must be in the future; omitted create deadline defaults to 30 days on the first accepted attempt. Only expired, non-removed source posts in the caller company may be reposted. Copies the current detail into a new PS3 post, retaining the source featured flag and charging the matching quota. Original post is unchanged; new post requires moderation. Replays do not reread mutable source content.",
        "x-jobfind-permission": "job:manage",
        "x-jobfind-internal": false,
        "x-jobfind-company-required": true,
        "x-jobfind-reject-unknown-query": true,
        "security": [
          {
            "internalSecret": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]{0,15}$",
              "format": "jobfind-id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 128,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
            },
            "description": "Required. Same user + key + canonical input returns the original accepted result; changed input or operation returns 409. Reuse for retries; a new key requests new work."
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]*$"
            },
            "description": "Authenticated identity forwarded by the trusted gateway only."
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "ADMIN",
                "COMPANY",
                "EMPLOYER"
              ]
            }
          },
          {
            "name": "x-company-id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[1-9][0-9]*$"
            },
            "description": "Required for non-admin callers; derived from the current account/company store by the gateway."
          },
          {
            "name": "x-company-status",
            "in": "header",
            "required": false,
            "schema": {
              "const": "S1"
            },
            "description": "Required for non-admin callers; derived from the current account/company store by the gateway."
          },
          {
            "name": "x-company-censor",
            "in": "header",
            "required": false,
            "schema": {
              "const": "CS1"
            },
            "description": "Required for non-admin callers; derived from the current account/company store by the gateway."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobRepost"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "headers": {
              "X-Correlation-ID": {
                "description": "Request correlation identifier",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "errCode": {
                      "const": 0
                    },
                    "data": {
                      "$ref": "#/components/schemas/Job"
                    },
                    "count": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "took": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 9007199254740991
                    },
                    "emailQueued": {
                      "type": "boolean"
                    },
                    "errMessage": {
                      "type": "string",
                      "maxLength": 1000
                    }
                  },
                  "required": [
                    "errCode",
                    "data"
                  ],
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "415": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "502": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "504": {
            "description": "Request rejected or dependency unavailable. errCode is application-specific; inspect HTTP status first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "internalSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "x-internal-secret",
        "description": "Private service credential. User context is supplied by the trusted gateway via x-user-id, x-user-role and x-company-id; never trust client-supplied identity headers."
      }
    },
    "schemas": {
...
      "JobRepost": {
        "type": "object",
        "properties": {
          "expectedRevision": {
            "type": "string",
            "pattern": "^jv1-[a-f0-9]{64}$",
            "minLength": 68,
            "maxLength": 68
          },
          "timeEnd": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1,
                "maximum": 8640000000000000
              },
              {
                "type": "string",
                "pattern": "^[1-9][0-9]{0,15}$",
                "format": "jobfind-id"
              }
            ]
          }
        },
        "required": [
          "timeEnd"
        ],
        "additionalProperties": false
      },
      "JobCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "pattern": "\\S"
          },
          "descriptionHTML": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "pattern": "\\S"
          },
          "descriptionMarkdown": {
            "type": "string",
            "maxLength": 200000
          },
          "categoryJobCode": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "\\S"
          },
          "addressCode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ]
          },
          "salaryJobCode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ]
          },
          "amount": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1,
                "maximum": 100000
              },
              {
                "type": "string",
                "pattern": "^(100000|[1-9][0-9]{0,4})$"
              }
            ]
          },
          "categoryJoblevelCode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ]
          },
          "categoryWorktypeCode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ]
          },
          "experienceJobCode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ]
          },
          "genderPostCode": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ]
          },
          "timeEnd": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1,
                "maximum": 8640000000000000
              },
              {
                "type": "string",
                "pattern": "^[1-9][0-9]{0,15}$",
                "format": "jobfind-id"
              }
            ]
          },
          "isHot": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "integer",
                "enum": [
                  0,
                  1
                ]
              }
            ]
          }
        },
        "required": [
          "name",
          "descriptionHTML",
          "categoryJobCode"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "errCode": {
            "type": "integer"
          },
          "errMessage": {
            "type": "string",
            "maxLength": 1000
          },
          "requestId": {
            "type": "string",
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
          },
          "conflict": {
            "type": "boolean"
          }
        },
        "required": [
          "errCode"
        ],
        "additionalProperties": true
      },
...
    }
  }
}